requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
marshmallow==2.19.2
pkg-resources==0.0.0
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.3
urllib3==1.25.3
Werkzeug==0.15.4
